+++ src/main/java/org/cloudwarp/probablychests/registry/PCEntities.java
@@ -15,27 +15,27 @@ public class PCEntities {
 	public static final EntityType<PCChestMimic> NORMAL_CHEST_MIMIC = Registry.register(
 			Registry.ENTITY_TYPE,
 			new Identifier(ProbablyChests.MOD_ID, "normal_chest_mimic"),
-			FabricEntityTypeBuilder.create(SpawnGroup.CREATURE, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
+			FabricEntityTypeBuilder.create(SpawnGroup.MONSTER, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
 	);
 	public static final EntityType<PCChestMimic> LUSH_CHEST_MIMIC = Registry.register(
 			Registry.ENTITY_TYPE,
 			new Identifier(ProbablyChests.MOD_ID, "lush_chest_mimic"),
-			FabricEntityTypeBuilder.create(SpawnGroup.CREATURE, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
+			FabricEntityTypeBuilder.create(SpawnGroup.MONSTER, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
 	);
 	public static final EntityType<PCChestMimic> ROCKY_CHEST_MIMIC = Registry.register(
 			Registry.ENTITY_TYPE,
 			new Identifier(ProbablyChests.MOD_ID, "rocky_chest_mimic"),
-			FabricEntityTypeBuilder.create(SpawnGroup.CREATURE, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
+			FabricEntityTypeBuilder.create(SpawnGroup.MONSTER, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
 	);
 	public static final EntityType<PCChestMimic> STONE_CHEST_MIMIC = Registry.register(
 			Registry.ENTITY_TYPE,
 			new Identifier(ProbablyChests.MOD_ID, "stone_chest_mimic"),
-			FabricEntityTypeBuilder.create(SpawnGroup.CREATURE, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
+			FabricEntityTypeBuilder.create(SpawnGroup.MONSTER, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
 	);
 	public static final EntityType<PCChestMimic> GOLD_CHEST_MIMIC = Registry.register(
 			Registry.ENTITY_TYPE,
 			new Identifier(ProbablyChests.MOD_ID, "gold_chest_mimic"),
-			FabricEntityTypeBuilder.create(SpawnGroup.CREATURE, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
+			FabricEntityTypeBuilder.create(SpawnGroup.MONSTER, PCChestMimic::new).dimensions(EntityDimensions.fixed(0.9f, 0.9f)).build()
 	);
 	//---------------------------------------
 	public static final EntityType<PCChestMimicPet> NORMAL_CHEST_MIMIC_PET = Registry.register(