+++ changelog.md
@@ -45,3 +45,19 @@ All notable changes to this project will be documented in this file.
 - When a chest becomes a mimic or pet mimic its inventory will be stored in the entity instead of dropping
 ### Fixed
 - Fixed crash when setting any spawn chance in the config to 0
+
+## [0.4.0] - 5/2/2022
+### Added
+- Mimics now spawn naturally
+- Config option to have easier mimics
+- Config option for naturally spawning mimics
+- Config option for natural mimic spawn rate
+### Changed
+- Decreased chance of mimic when opening chest
+- Mimics no longer drown
+- Increased chance to find mimic key fragments in chests
+- Mimics can now drop key fragments
+- Pet mimics now have different animation for sitting and standing
+### Fixed
+- Fixed issue where opening a mimic with a bow would cause bow to fire when the gui was closed
+- Fixed pet mimics were not allowed in peaceful